flatten types
