--- src/NServiceBus.AzureFunctions.StorageQueues/Serverless/ServerlessEndpoint.cs
@@ -1,113 +1,0 @@
-﻿namespace NServiceBus.AzureFunctions.StorageQueues
-{
-    using System;
-    using System.IO;
-    using System.Reflection;
-    using System.Runtime.Loader;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using Logging;
-
-    /// <summary>
-    /// An NServiceBus endpoint which does not receive messages automatically but only handles messages explicitly passed to it
-    /// by the caller.
-    /// Instances of <see cref="ServerlessEndpoint{TConfiguration}" /> can be cached and are thread-safe.
-    /// </summary>
-    public abstract class ServerlessEndpoint<TConfiguration>
-        where TConfiguration : ServerlessEndpointConfiguration
-    {
-        /// <summary>
-        /// Allows to forcefully initialize the endpoint if it hasn't been initialized yet.
-        /// </summary>
-        /// <param name="executionContext">The execution context.</param>
-        /// <param name="token">The cancellation token or default cancellation token.</param>
-        protected async Task InitializeEndpointIfNecessary(FunctionExecutionContext executionContext, CancellationToken token = default)
-        {
-            if (pipeline == null)
-            {
-                await semaphoreLock.WaitAsync(token).ConfigureAwait(false);
-                try
-                {
-                    if (pipeline == null)
-                    {
-                        LogManager.GetLogger("Previews").Info("NServiceBus.AzureFunctions.ServiceBus is a preview package. Preview packages are licensed separately from the rest of the Particular Software platform and have different support guarantees. You can view the license at https://particular.net/eula/previews and the support policy at https://docs.particular.net/previews/support-policy. Customer adoption drives whether NServiceBus.AzureFunctions.ServiceBus will be incorporated into the Particular Software platform. Let us know you are using it, if you haven't already, by emailing us at support@particular.net.");
-                        var endpoint = await endpointFactory(executionContext).ConfigureAwait(false);
-
-                        pipeline = configuration.PipelineInvoker;
-                    }
-                }
-                finally
-                {
-                    semaphoreLock.Release();
-                }
-            }
-        }
-
-        internal static void LoadAssemblies(string assemblyDirectory)
-        {
-            var binFiles = Directory.EnumerateFiles(
-                assemblyDirectory,
-                "*.dll",
-                SearchOption.TopDirectoryOnly);
-
-            var assemblyLoadContext = AssemblyLoadContext.GetLoadContext(Assembly.GetExecutingAssembly());
-            foreach (var binFile in binFiles)
-            {
-                try
-                {
-                    var assemblyName = AssemblyName.GetAssemblyName(binFile);
-                    if (IsRuntimeAssembly(assemblyName.GetPublicKeyToken()))
-                    {
-                        continue;
-                    }
-
-                    // LoadFromAssemblyName works when actually running inside a function as FunctionAssemblyLoadContext probes the "bin" folder for the assembly name
-                    // this doesn't work when running with a different AssemblyLoadContext (e.g. tests) and the assembly needs to be loaded by the full path instead.
-                    assemblyLoadContext.LoadFromAssemblyPath(binFile);
-                    //assemblyLoadContext.LoadFromAssemblyName(assemblyName);
-                }
-                catch (Exception e)
-                {
-                    LogManager.GetLogger<FunctionEndpoint>().DebugFormat(
-                        "Failed to load assembly {0}. This error can be ignored if the assembly isn't required to execute the function.{1}{2}",
-                        binFile, Environment.NewLine, e);
-                }
-            }
-        }
-
-        static bool IsRuntimeAssembly(byte[] publicKeyToken)
-        {
-            var tokenString = BitConverter.ToString(publicKeyToken).Replace("-", string.Empty).ToLowerInvariant();
-
-            switch (tokenString)
-            {
-                case "b77a5c561934e089": // Microsoft
-                case "7cec85d7bea7798e":
-                case "b03f5f7f11d50a3a":
-                case "31bf3856ad364e35":
-                case "cc7b13ffcd2ddd51":
-                case "adb9793829ddae60":
-                case "7e34167dcc6d6d8c": // Microsoft.Azure.ServiceBus
-                case "23ec7fc2d6eaa4a5": // Microsoft.Data.SqlClient
-                case "50cebf1cceb9d05e": // Mono.Cecil
-                case "30ad4fe6b2a6aeed": // Newtonsoft.Json
-                case "9fc386479f8a226c": // NServiceBus
-                    return true;
-                default:
-                    return false;
-            }
-        }
-
-        /// <summary>
-        /// Provides a function to locate the file system directory containing the binaries to be loaded and scanned.
-        /// When using functions, assemblies are moved to a 'bin' folder within ExecutionContext.FunctionAppDirectory.
-        /// </summary>
-        protected Func<FunctionExecutionContext, string> AssemblyDirectoryResolver = functionExecutionContext => Path.Combine(functionExecutionContext.ExecutionContext.FunctionAppDirectory, "bin");
-
-        readonly SemaphoreSlim semaphoreLock = new SemaphoreSlim(initialCount: 1, maxCount: 1);
-        private protected Func<FunctionExecutionContext, Task<IEndpointInstance>> endpointFactory;
-        private protected StorageQueueTriggeredEndpointConfiguration configuration;
-
-        private protected PipelineInvoker pipeline;
-    }
-}
